+++ apps/enmeshed/lib/account/mailbox/mailbox_view.dart
@@ -6,20 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 
 import '/core/core.dart';
+import 'mailbox_filter_option.dart';
 import 'modals/select_contact_filter.dart';
-
-enum MailboxFilterOption {
-  incoming(Icons.mail, Icons.mail_outline),
-  actionRequired(Icons.chat_bubble, Icons.chat_bubble_outline),
-  unread(Icons.mark_email_unread, Icons.mark_email_unread_outlined),
-  withAttachment(Icons.attachment, Icons.attachment_outlined),
-  outgoing(Icons.send, Icons.send_outlined);
-
-  final IconData filterIcon;
-  final IconData emptyListIcon;
-
-  const MailboxFilterOption(this.filterIcon, this.emptyListIcon);
-}
 
 class MailboxView extends StatefulWidget {
   final String accountId;
